Use `screen` instead of `innerWidth/Height`
In the `getMobileDimensions` method, we're calculating the the width using
`window.innerWidth`, which in some edge cases can actually be larger than the
screen.  Screen will return the width and height of the actual device screen,
which I think is more in line with what we want in this case.

DEV-1643