Use `absolute` instead of `fixed` for mobile
Using position `fixed` fixes an element relative to the viewport.  In our case
we were fixing the wrapper and the iframe and setting hard-coded widths on them.
Position fixed combined with a width that's larger than the viewport will result
in unreachable content.  You can't scroll an element when it's fixed.  Because
we've (hopefully) fixed the incorrect width issue with teh previous commit, this
shouldn't really be necessary but I think it's safer to always allow users to
scroll the content as a fallback.

DEV-1643