Mobile viewport (#26)
* Use `screen` instead of `innerWidth/Height`

In the `getMobileDimensions` method, we're calculating the the width using
`window.innerWidth`, which in some edge cases can actually be larger than the
screen.  Screen will return the width and height of the actual device screen,
which I think is more in line with what we want in this case.

DEV-1643

* Use `absolute` instead of `fixed` for mobile

Using position `fixed` fixes an element relative to the viewport.  In our case
we were fixing the wrapper and the iframe and setting hard-coded widths on them.
Position fixed combined with a width that's larger than the viewport will result
in unreachable content.  You can't scroll an element when it's fixed.  Because
we've (hopefully) fixed the incorrect width issue with teh previous commit, this
shouldn't really be necessary but I think it's safer to always allow users to
scroll the content as a fallback.

DEV-1643

* Bump version

* Use window dimensions for checking orientation

Screen dimensions are constant dimensions for the device screen that don't
change when the orientation changes.  So we also need to track the window
dimensions to see if orientation has changed.