Make sure iFrame has pixel width for containers
When dealing with iFrames in iOS, the `width` rule is apparently ignored if it
doesn't have a pixel value.  This means that the iframe content will be expanded
to the full width of whatever the content is.  In some cases, the iframe content
is bigger than the device screen, so the iframe content overflows outside of the
viewport.  This coupled with user-disabled scrolling results in the user not
being able to get to the actual content.