Remove scrollTo call on scroll listener
We had a `scrollTo` call set that scrolled the window to the top for _every_
scroll event.  This effectively prevented users from scrolling at all, which is
bad in cases where our iframe/wrapper dimensions are off incorrect and content
is outside of the viewport.  I believe this was in place to make sure the
iframe/wrapper divs were always visible on the screen.  Setting the wrapper to
'position: fixed` should accomplish that while still allowing the content to be
scrollable.