Adjust wrapper styles
- The wrapper position was previously changed to `absolute` thinking that it would
solve the problem of not being able to scroll to offscreen content.  It turns
out there were other areas of the codebase that were preventing this, not the
positioning.  Changing to `fixed` will keep the wrapper in the correct place on
the screen.  Additional rules were added to scroll overflow content.
- We need to adjust the wrapper width on resize as well as the iframe width