Revert to absolute position for the wrapper (#34)
* Revert to absolute position for the wrapper

In a previous commit we changed the wrapper positioning to `fixed` which
fixed one problem and caused a bunch of other problems.  I also removed
the scrol listener because it was preventing users from scrolling to
content.  It turns out that other things in this experience rely on that
scroll listener in place, so reverting that change seems to make the
most sense.

[BUGREP-1259]

* Bump version