feat: Avoid defaulting to English if locale is not set
If `locale" is not specified, we prefer to honor an Embedded user's default browser language setting (if it is supported by our web app) instead of always forcing users into viewing our web app in English. Our backend now allows the `user_culture` param to be optional, and if it is blank or omitted then it will tell the web app to prefer the user's default browser language.

Also in this feat:
- Update tests
- Simplify test mock GUIDs
- Correct a few comments